+++ sdk/dotnet/Inputs/ApplicationAppRoleArgs.cs
@@ -37,7 +37,7 @@ namespace Pulumi.AzureAD.Inputs
         public Input<string> DisplayName { get; set; } = null!;
 
         /// <summary>
-        /// The unique identifier of the `app_role`.
+        /// The unique identifier of the permision. This attribute is computed and cannot be specified manually in this block. If you need to specify a custom `id`, it's recommended to use the azuread.ApplicationOAuth2Permission resource.
         /// </summary>
         [Input("id")]
         public Input<string>? Id { get; set; }