+++ sdk/go/azuread/applicationPassword.go
@@ -57,8 +57,6 @@ import (
 type ApplicationPassword struct {
 	pulumi.CustomResourceState
 
-	// Deprecated: Deprecated in favour of `application_object_id` to prevent confusion
-	ApplicationId pulumi.StringOutput `pulumi:"applicationId"`
 	// The Object ID of the Application for which this password should be created. Changing this field forces a new resource to be created.
 	ApplicationObjectId pulumi.StringOutput `pulumi:"applicationObjectId"`
 	// A description for the Password.
@@ -82,6 +80,9 @@ func NewApplicationPassword(ctx *pulumi.Context,
 		return nil, errors.New("missing one or more required arguments")
 	}
 
+	if args.ApplicationObjectId == nil {
+		return nil, errors.New("invalid value for required argument 'ApplicationObjectId'")
+	}
 	if args.Value == nil {
 		return nil, errors.New("invalid value for required argument 'Value'")
 	}
@@ -107,8 +108,6 @@ func GetApplicationPassword(ctx *pulumi.Context,
 
 // Input properties used for looking up and filtering ApplicationPassword resources.
 type applicationPasswordState struct {
-	// Deprecated: Deprecated in favour of `application_object_id` to prevent confusion
-	ApplicationId *string `pulumi:"applicationId"`
 	// The Object ID of the Application for which this password should be created. Changing this field forces a new resource to be created.
 	ApplicationObjectId *string `pulumi:"applicationObjectId"`
 	// A description for the Password.
@@ -126,8 +125,6 @@ type applicationPasswordState struct {
 }
 
 type ApplicationPasswordState struct {
-	// Deprecated: Deprecated in favour of `application_object_id` to prevent confusion
-	ApplicationId pulumi.StringPtrInput
 	// The Object ID of the Application for which this password should be created. Changing this field forces a new resource to be created.
 	ApplicationObjectId pulumi.StringPtrInput
 	// A description for the Password.
@@ -149,10 +146,8 @@ func (ApplicationPasswordState) ElementType() reflect.Type {
 }
 
 type applicationPasswordArgs struct {
-	// Deprecated: Deprecated in favour of `application_object_id` to prevent confusion
-	ApplicationId *string `pulumi:"applicationId"`
 	// The Object ID of the Application for which this password should be created. Changing this field forces a new resource to be created.
-	ApplicationObjectId *string `pulumi:"applicationObjectId"`
+	ApplicationObjectId string `pulumi:"applicationObjectId"`
 	// A description for the Password.
 	Description *string `pulumi:"description"`
 	// The End Date which the Password is valid until, formatted as a RFC3339 date string (e.g. `2018-01-01T01:02:03Z`). Changing this field forces a new resource to be created.
@@ -169,10 +164,8 @@ type applicationPasswordArgs struct {
 
 // The set of arguments for constructing a ApplicationPassword resource.
 type ApplicationPasswordArgs struct {
-	// Deprecated: Deprecated in favour of `application_object_id` to prevent confusion
-	ApplicationId pulumi.StringPtrInput
 	// The Object ID of the Application for which this password should be created. Changing this field forces a new resource to be created.
-	ApplicationObjectId pulumi.StringPtrInput
+	ApplicationObjectId pulumi.StringInput
 	// A description for the Password.
 	Description pulumi.StringPtrInput
 	// The End Date which the Password is valid until, formatted as a RFC3339 date string (e.g. `2018-01-01T01:02:03Z`). Changing this field forces a new resource to be created.