+++ sdk/python/pulumi_azuread/_inputs.py
@@ -37,7 +37,7 @@ class ApplicationAppRoleArgs:
         :param pulumi.Input[Sequence[pulumi.Input[str]]] allowed_member_types: Specifies whether this app role definition can be assigned to users and groups by setting to `User`, or to other applications (that are accessing this application in daemon service scenarios) by setting to `Application`, or to both.
         :param pulumi.Input[str] description: Permission help text that appears in the admin app assignment and consent experiences.
         :param pulumi.Input[str] display_name: Display name for the permission that appears in the admin consent and app assignment experiences.
-        :param pulumi.Input[str] id: The unique identifier of the `app_role`.
+        :param pulumi.Input[str] id: The unique identifier of the permision. This attribute is computed and cannot be specified manually in this block. If you need to specify a custom `id`, it's recommended to use the ApplicationOAuth2Permission resource.
         :param pulumi.Input[bool] is_enabled: Determines if the permission is enabled: defaults to `true`.
         :param pulumi.Input[str] value: The value of the scope claim that the resource application should expect in the OAuth 2.0 access token.
         """
@@ -91,7 +91,7 @@ class ApplicationAppRoleArgs:
     @pulumi.getter
     def id(self) -> Optional[pulumi.Input[str]]:
         """
-        The unique identifier of the `app_role`.
+        The unique identifier of the permision. This attribute is computed and cannot be specified manually in this block. If you need to specify a custom `id`, it's recommended to use the ApplicationOAuth2Permission resource.
         """
         return pulumi.get(self, "id")